+++ llvm/docs/ReleaseNotes.md
@@ -238,6 +238,7 @@ Makes programs 10x faster by doing Special New Thing.
 * Bump Svukte extension to 1.0.
 * Remove experimental from Zicfiss.
 * Added support for `Sspmp`, `Sspmpen` and `Smpmpdeleg` extensions.
+* Bumped the supported version of the experimental `Zibi` extension from 0.1 to 0.7.
 
 ### Changes to the WebAssembly Backend